Add support for `action`, as a way to perform side-effects
This adds a new export, `action`, which can be used to perform
side-effects.

This is used during transitions and takes the `context` and the `event`,
just like reducers.

I also refactored transitions/immediates since that code was mostly the
same. The result of the refactor means this new feature only adds 2
bytes.